pyproject.toml
[tool.poetry]
name = "colrag"
version = "0.1.0"
description = "A RAG pipeline using ColBERT via RAGatouille"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "colrag"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.0"
python-docx = "^0.8.11"
beautifulsoup4 = "^4.10.0"
PyPDF2 = "^2.0.0"
openpyxl = "^3.0.7"
ragatouille = "0.0.8.post2"
tqdm = "^4.62.3"
python-dotenv = "^0.19.0"
pydantic = "^1.8.2"
typer = "^0.4.0"
rich = "^10.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
colrag = "colrag.cli:app"